add Notification status CRD and controller logic (#512)
This api surfaces the status of the notification controller, which
informs users on the status of:
- which notifications have been delivered
- whether any errors occurred
- whether any warnings occurred

This is intended to provide a better UX for understanding the state of
the Notification configuration. The lifecycle of the Notification CR is
managed by reconciler-manager, and the status is updated by the
notification container in the reconciler.